dddd

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -8,7 +8,8 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-    size_t i;
+    unsigned int i;
+
     if (array == NULL || action == NULL)
         return;
